Reorganize BaseController. DesktopController is now a IConcretController. Add bindings to the PlayerInputAction.
