+++ UMI3D-Browser-Desktop/Assets/Dependencies/UMI3D Screen Browsers Base/Scripts/Controllers/BaseController_InteractionsAndParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b033f64da26f1de45994be7d47f4d347
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
