+++ UMI3D-Browser-Desktop/Assets/Dependencies/UMI3D Screen Browsers Base/Scripts/Controllers/CursorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee3bd1ab134651647b67414df22ae2f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
