+++ UMI3D-Browser-Desktop/Assets/Dependencies/UMI3D Screen Browsers Base/Scripts/Controllers/IConcreteController.cs
@@ -1,0 +1,32 @@
+/*
+Copyright 2019 - 2023 Inetum
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+using System.Collections;
+using System.Collections.Generic;
+using umi3d.cdk.interaction;
+using umi3d.common.interaction;
+using UnityEngine;
+
+namespace umi3d.baseBrowser.Controller
+{
+    public interface IConcreteController
+    {
+        List<AbstractUMI3DInput> Inputs { get; }
+
+        void Awake();
+        void Update();
+        AbstractUMI3DInput FindInput(EventDto evt, bool unused = true, bool tryToFindInputForHoldableEvent = false);
+    }
+}
