+++ UMI3D-Browser-Desktop/Assets/Dependencies/UMI3D Screen Browsers Base/Scripts/Controllers/IConcreteController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbca7b299cb93374a842e97a6e443e81
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
