+++ UMI3D-Browser-Desktop/Assets/Project/Scenes/Environment.unity
@@ -333,6 +333,7 @@ GameObject:
   m_Component:
   - component: {fileID: 200932040}
   - component: {fileID: 200932041}
+  - component: {fileID: 200932042}
   m_Layer: 0
   m_Name: Controllers
   m_TagString: Untagged
@@ -388,6 +389,25 @@ MonoBehaviour:
   m_DefaultActionMap: Player
   m_SplitScreenIndex: -1
   m_Camera: {fileID: 0}
+--- !u!114 &200932042
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 200932039}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: eef9afd125ba6a146a1f78de411a6729, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  projectionMemory: {fileID: 498482125}
+  ObjectMenu: {fileID: 0}
+  CameraTransform: {fileID: 907048306}
+  InteractionMapper: {fileID: 498482124}
+  dofGroups: 04000000050000000600000001000000020000000300000008000000090000000a000000
+  interactionBoneType: 58
+  hoverBoneType: 58
 --- !u!1 &395169811
 GameObject:
   m_ObjectHideFlags: 0
@@ -536,7 +556,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   Controllers:
-  - {fileID: 2096641628}
+  - {fileID: 0}
   shouldProjectHoldableEventOnSpecificInput: 0
 --- !u!114 &498482125
 MonoBehaviour:
@@ -674,7 +694,7 @@ PrefabInstance:
     - target: {fileID: 9168339693674874675, guid: da6d27ee883946e46a8d6c30cad5b8a6,
         type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 9168339693674874675, guid: da6d27ee883946e46a8d6c30cad5b8a6,
         type: 3}
@@ -1356,7 +1376,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2096641627}
-  - component: {fileID: 2096641628}
   m_Layer: 0
   m_Name: DesktopController
   m_TagString: Untagged
@@ -1381,24 +1400,3 @@ Transform:
   m_Father: {fileID: 200932040}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &2096641628
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2096641626}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 17b5e0768f3351d4a94af9701bffda78, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  projectionMemory: {fileID: 498482125}
-  ObjectMenu: {fileID: 0}
-  CameraTransform: {fileID: 907048306}
-  InteractionMapper: {fileID: 498482124}
-  dofGroups: 04000000050000000600000001000000020000000300000008000000090000000a000000
-  interactionBoneType: 58
-  hoverBoneType: 58
-  ManipulationActionInput:
-  - {fileID: 647256037}
